implemented drawEllipseByCoordinates to keep drawEllipse function parameters not touched
